# examples/fullstack-slop-lab/openapi.yaml
openapi: 3.1.0
info:
  title: NexusFlow AI API
  description: Intentionally awkward beta fixture for UIdetox.
  version: 0.1.0
paths:
...
  /api/projects:
    get:
      summary: List Projects
      operationId: list_projects_api_projects_get
      parameters:
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
      - name: search
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Search
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProjectResponse'
                title: Response List Projects Api Projects Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      summary: Create Project
      operationId: create_project_api_projects_post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/projects/{project_id}:
    get:
      summary: Get Project
      operationId: get_project_api_projects__project_id__get
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: integer
          title: Project Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectDetailResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      summary: Update Project
      operationId: update_project_api_projects__project_id__patch
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: integer
          title: Project Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      summary: Delete Project
      operationId: delete_project_api_projects__project_id__delete
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: integer
          title: Project Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/team/invite:
    post:
      summary: Invite Team Member
      operationId: invite_team_member_api_team_invite_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TeamInvite'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamMemberResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/team/{member_id}:
    delete:
      summary: Delete Team Member
      operationId: delete_team_member_api_team__member_id__delete
      parameters:
      - name: member_id
        in: path
        required: true
        schema:
          type: integer
          title: Member Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/settings:
    get:
      summary: Get Settings
      operationId: get_settings_api_settings_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkspaceSettingsResponse'
    put:
      summary: Save Settings
      operationId: save_settings_api_settings_put
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SettingsUpdate'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkspaceSettingsResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/workflows/{workflow_id}/pause:
    post:
      summary: Pause Workflow
      operationId: pause_workflow_api_workflows__workflow_id__pause_post
      parameters:
      - name: workflow_id
        in: path
        required: true
        schema:
          type: integer
          title: Workflow Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AutomationResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/notifications/{notification_id}/seen:
    post:
      summary: Mark Notification Seen
      operationId: mark_notification_seen_api_notifications__notification_id__seen_post
      parameters:
      - name: notification_id
        in: path
        required: true
        schema:
          type: integer
          title: Notification Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/experiments/{experiment_key}:
    put:
      summary: Save Experiment
      operationId: save_experiment_api_experiments__experiment_key__put
      parameters:
      - name: experiment_key
        in: path
        required: true
        schema:
          type: string
          title: Experiment Key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExperimentUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExperimentResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/accounts/{account_id}:
    patch:
      summary: Update Account Health
      operationId: update_account_health_api_accounts__account_id__patch
      parameters:
      - name: account_id
        in: path
        required: true
        schema:
          type: integer
          title: Account Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountHealthUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/data-sources/{source_id}/sync:
    post:
      summary: Sync Data Source
      operationId: sync_data_source_api_data_sources__source_id__sync_post
      parameters:
      - name: source_id
        in: path
        required: true
        schema:
          type: integer
          title: Source Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataSourceResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/governance/approvals/{approval_id}/decision:
    post:
      summary: Decide Governance Approval
      operationId: decide_governance_approval_api_governance_approvals__approval_id__decision_post
      parameters:
      - name: approval_id
        in: path
        required: true
        schema:
          type: integer
          title: Approval Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DecisionUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApprovalResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/journeys/{journey_id}/publish:
    post:
      summary: Publish Journey
      operationId: publish_journey_api_journeys__journey_id__publish_post
      parameters:
      - name: journey_id
        in: path
        required: true
        schema:
          type: integer
          title: Journey Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JourneyResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ProjectCreate:
      properties:
        name:
          type: string
          minLength: 2
          title: Name
        description:
          type: string
          title: Description
          default: ''
        status:
          type: string
          enum:
          - planning
          - active
          - at-risk
          - completed
          title: Status
          default: planning
        progress:
          type: integer
          maximum: 100.0
          minimum: 0.0
          title: Progress
          default: 0
        budget:
          type: number
          minimum: 0.0
          title: Budget
          default: 0
        due_date:
          anyOf:
          - type: string
          - type: 'null'
          title: Due Date
        owner_name:
          type: string
          title: Owner Name
          default: Mara Voss
        tags:
          items:
            type: string
          type: array
          title: Tags
      additionalProperties: false
      type: object
      required:
      - name
      title: ProjectCreate
    ProjectDetailResponse:
      properties:
        id:
          type: integer
          title: Id
        name:
          type: string
          title: Name
        description:
          type: string
          title: Description
        status:
          type: string
          enum:
          - planning
          - active
          - at-risk
          - completed
          title: Status
        progress:
          type: integer
          title: Progress
        budget:
          type: number
          title: Budget
        due_date:
          anyOf:
          - type: string
          - type: 'null'
          title: Due Date
        owner_name:
          type: string
          title: Owner Name
        tags:
          items:
            type: string
          type: array
          title: Tags
        created_at:
          type: string
          title: Created At
        activity:
          items:
            $ref: '#/components/schemas/ActivityResponse'
          type: array
          title: Activity
      type: object
      required:
      - id
      - name
      - description
      - status
      - progress
      - budget
      - due_date
      - owner_name
      - tags
      - created_at
      - activity
      title: ProjectDetailResponse
    ProjectResponse:
      properties:
        id:
          type: integer
          title: Id
        name:
          type: string
          title: Name
        description:
          type: string
          title: Description
        status:
          type: string
          enum:
          - planning
          - active
          - at-risk
          - completed
          title: Status
        progress:
          type: integer
          title: Progress
        budget:
          type: number
          title: Budget
        due_date:
          anyOf:
          - type: string
          - type: 'null'
          title: Due Date
        owner_name:
          type: string
          title: Owner Name
        tags:
          items:
            type: string
          type: array
          title: Tags
        created_at:
          type: string
          title: Created At
      type: object
      required:
      - id
      - name
      - description
      - status
      - progress
      - budget
      - due_date
      - owner_name
      - tags
      - created_at
      title: ProjectResponse
...